requirements.txt
django
django-crispy-forms==1.9.2
django-weasyprint

djangorestframework

python-slugify